requirements.txt
bidict==0.22.1
blinker==1.6.2
chess==1.9.4
click==8.1.3
colorama==0.4.6
Flask==2.3.2
Flask-SocketIO==5.3.4
greenlet==2.0.2
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.33
protobuf==3.20.3
python-chess==1.999
python-engineio==4.4.1
python-socketio==5.8.0
typing-extensions==4.5.0
Werkzeug==2.3.4
zipp==3.15.0
